models/predict_pue/requirements_model_pue.txt
pandas
numpy
requests
pyyaml
fastapi
uvicorn[standard]
pydantic

